--- src/sortable-list/components/index.ts
@@ -1,1 +1,0 @@
-export * from './sortable-list.component';
