+++ src/sortable/models/draggable-item.ts
